requirements.txt
aiohttp==3.7.3
aiohttp-cors==0.7.0
async-timeout==3.0.1
attrs==20.3.0
boto3==1.16.26
botocore==1.19.26
chardet==3.0.4
click==7.1.2
colorclass==2.2.0
colorlog==4.6.2
croniter==0.3.36
faust==1.10.4
idna==2.10
jmespath==0.10.0
kafka-python==1.4.7
mode==4.3.2
multidict==5.0.2
mypy-extensions==0.4.3
natsort==7.1.0
numpy==1.19.4
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
opentracing==1.3.0
Pillow==8.0.1
python-dateutil==2.8.1
robinhood-aiokafka==1.1.6
s3transfer==0.3.3
six==1.15.0
terminaltables==3.1.0
tqdm==4.54.0
typing-extensions==3.7.4.3
urllib3==1.26.2
venusian==1.2.0
yarl==1.6.3
